backend/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
httpx==0.27.0
beautifulsoup4==4.12.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
google-cloud-firestore>=2.11.1
firebase-admin>=6.2.0
python-dotenv>=1.0.0
email-validator>=2.0.0
stripe>=5.0.0
google-genai
